requirements.txt
openai
tqdm
fpdf==1.7.2
colorama
python-dotenv
